Rename InputItem files to InputValue
