+++ source/uwp/Renderer/AdaptiveCardRenderer.vcxproj.filters
@@ -18,7 +18,6 @@
     <ClCompile Include="lib\AdaptiveShowCardAction.cpp" />
     <ClCompile Include="lib\AdaptiveSubmitAction.cpp" />
     <ClCompile Include="lib\AdaptiveActionEventArgs.cpp" />
-    <ClCompile Include="lib\InputItem.cpp" />
     <ClCompile Include="lib\AdaptiveChoiceInput.cpp" />
     <ClCompile Include="lib\AdaptiveChoiceSetInput.cpp" />
     <ClCompile Include="lib\AdaptiveDateInput.cpp" />
@@ -76,6 +75,7 @@
     <ClCompile Include="lib\AdaptiveError.cpp" />
     <ClCompile Include="lib\AdaptiveWarning.cpp" />
     <ClCompile Include="lib\HtmlHelpers.cpp" />
+    <ClCompile Include="lib\InputValue.cpp" />
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="lib\AdaptiveCard.h" />
@@ -102,7 +102,6 @@
     <ClInclude Include="lib\AdaptiveShowCardAction.h" />
     <ClInclude Include="lib\AdaptiveSubmitAction.h" />
     <ClInclude Include="lib\AdaptiveActionEventArgs.h" />
-    <ClInclude Include="lib\InputItem.h" />
     <ClInclude Include="lib\AdaptiveChoiceInput.h" />
     <ClInclude Include="lib\AdaptiveChoiceSetInput.h" />
     <ClInclude Include="lib\AdaptiveDateInput.h" />
@@ -160,6 +159,7 @@
     <ClInclude Include="lib\AdaptiveError.h" />
     <ClInclude Include="lib\AdaptiveWarning.h" />
     <ClInclude Include="lib\HtmlHelpers.h" />
+    <ClInclude Include="lib\InputValue.h" />
   </ItemGroup>
   <ItemGroup>
     <Midl Include="idl\AdaptiveCards.Rendering.Uwp.idl" />
