+++ source/uwp/Renderer/lib/InputValue.cpp
@@ -1,5 +1,5 @@
 #include "pch.h"
-#include "InputItem.h"
+#include "InputValue.h"
 #include "json/json.h"
 #include "XamlHelpers.h"
 #include <windows.globalization.datetimeformatting.h>
