+++ source/uwp/Renderer/lib/InputValue.h
