+++ source/uwp/Renderer/lib/Util.h
@@ -4,7 +4,7 @@
 #include <string>
 
 #include "AdaptiveCards.Rendering.Uwp.h"
-#include "InputItem.h"
+#include "InputValue.h"
 #include <BaseCardElement.h>
 #include <BaseActionElement.h>
 #include <ChoiceInput.h>
